IGNITE-10594 Remove specific JSR310 handling, add java.time test to regular suite.
